requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.29
pymysql==1.1.0
python-dotenv==1.0.1

jinja2==3.1.3
python-multipart==0.0.9

passlib[bcrypt]==1.7.4

scikit-learn==1.4.2
numpy==1.26.4
rapidfuzz==3.6.2

email-validator==2.1.1

itsdangerous==2.1.2